dev-java/openpdf: Changed source and added patch
There are a bunch of resources that needed to be added in, which
weren't from the maven sources. I had to use main upstream DL
and use the addres function.

The patch adds some extra constructors for itextpdf compatibility
and adds public to the main one since there were errors about it
being private.